creation des tables produits commandes clients

+++ db/migrate/20150527090700_create_produit.rb
@@ -1,11 +1,11 @@
 class CreateProduit < ActiveRecord::Migration
   def change
-  	create_table :produits do |c|
+  	  	create_table :produits do |c|
   		c.string :refproduit
   		c.string :article
   		c.string :descriptif
   		c.integer :stock
-  		c.integer :prixproduit
+  		c.float :prixproduit
   	end
   end
 end
